requirements/base.txt
django==4.2.2
psycopg2-binary==2.9.6
python-environ==0.4.54
django-cors-headers==4.1.0
django-jazzmin==2.6.0
drf-yasg==1.21.6
djangorestframework==3.14.0
django-filter==23.2
coreapi==2.3.3
djangorestframework-simplejwt==5.2.2
django-solo==2.1.0
sorl-thumbnail==12.9.0
Pillow==10.0.0
